Restruxture movies and add support for complex qualities, resolutions, sources, codes, root dirs, and history. Needs work.
