+++ sickchill/movies/__init__.py
@@ -1,0 +1,2 @@
+from .movies_list import MoviesList
+from .models import *
