+++ sickchill/movies/templates/movies/remove.mako
